% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_column_names.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Cleans and standarizes column names for the loaded data.}
\usage{
clean_column_names(data)
}
\arguments{
\item{data}{a dataframe with the uncleaned columns.}
}
\value{
a dataframe with the cleaned columns.
}
\description{
Cleans and standarizes column names for the loaded data.
}
\examples{
df_raw <- data.frame(V1 = 1:3, V2 = 1:3, V3 = 1:3, V4 = 1:3,
                     V5 = 1:3, V6 = 1:3, V7 = 1:3)
clean_column_names(df_raw)
}
