% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test_accuracy.R
\name{compute_test_accuracy}
\alias{compute_test_accuracy}
\title{Compute Test Accuracy}
\usage{
compute_test_accuracy(predicted, actual)
}
\arguments{
\item{predicted}{A vector of predicted class labels.}

\item{actual}{A vector of true class labels.}
}
\value{
A numeric value between 0 and 1 representing the test accuracy.
}
\description{
Calculates the proportion of correct predictions.
}
\examples{
predicted <- c("yes", "no", "yes", "yes")
actual <- c("yes", "no", "no", "yes")
compute_test_accuracy(predicted, actual)
# Returns 0.75
}
