% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_levels.R
\name{encode_levels}
\alias{encode_levels}
\title{Encode categorical variables with predefined levels.}
\usage{
encode_levels(x)
}
\arguments{
\item{x}{A character vector to encode.}
}
\value{
A numeric vector representing encoded values.
}
\description{
Encode categorical variables with predefined levels.
}
\examples{
encode_levels(c("low", "med", "high", "vhigh", "acc", "vgood", "3"))
# Returns: 1, 2, 3, 4, 2, 4, 3
}
