% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_safety.R
\name{encode_safety}
\alias{encode_safety}
\title{Encode the safety variable.}
\usage{
encode_safety(x)
}
\arguments{
\item{x}{A character vector representing safety levels.}
}
\value{
A numeric vector representing the safety encoding.
}
\description{
Encode the safety variable.
}
\examples{
encode_safety(c("low", "med", "high", "2"))
# Returns: 1, 2, 3, 2

}
