% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_car_data.R
\name{load_car_data}
\alias{load_car_data}
\title{Load and clean the Car Evaluation dataset from UCI}
\usage{
load_car_data(output_path)
}
\arguments{
\item{output_path}{A string specifying the file path to save the cleaned CSV dataset.}
}
\value{
A tibble containing the cleaned and labeled dataset.
}
\description{
Downloads the Car Evaluation dataset from the UCI Machine Learning Repository,
assigns column names, cleans them, and saves the result as a CSV file.
}
\examples{
\dontrun{
cleaned_data <- load_car_data(output_path = "data/original/car_data.csv")
}
}
