% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conf_matrix.R
\name{plot_conf_matrix}
\alias{plot_conf_matrix}
\title{Plot Confusion Matrix}
\usage{
plot_conf_matrix(true_labels, predicted_labels, title = "Confusion Matrix")
}
\arguments{
\item{true_labels}{A factor vector of actual labels.}

\item{predicted_labels}{A factor vector of predicted labels.}

\item{title}{Optional title for the plot.}
}
\value{
A ggplot object representing the confusion matrix heatmap.
}
\description{
Generates a heatmap of the confusion matrix from actual and predicted labels.
}
\examples{
true <- factor(c("yes", "no", "yes", "no", "yes"))
predicted <- factor(c("yes", "no", "no", "no", "yes"))
plot_conf_matrix(true, predicted)
}
