% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_final_model.R
\name{train_final_model}
\alias{train_final_model}
\title{Train final kNN model}
\usage{
train_final_model(train_x, train_y, best_k)
}
\arguments{
\item{train_x}{Training predictors.}

\item{train_y}{Training response.}

\item{best_k}{Optimal k value.}
}
\value{
Trained kNN model.
}
\description{
Train final kNN model
}
\examples{
df <- data.frame(x1 = rnorm(50), x2 = rnorm(50))
y <- factor(sample(c("yes", "no"), 50, replace = TRUE))
model <- train_final_model(df, y, best_k = 5)
}
